Add GH GaugeWave executable
