+++ src/Evolution/Executables/GeneralizedHarmonic/EvolveGeneralizedHarmonic.hpp
@@ -3,413 +3,51 @@
 
 #pragma once
 
-#include <cstddef>
 #include <cstdint>
 #include <vector>
 
-#include "ApparentHorizons/ComputeItems.hpp"
-#include "ApparentHorizons/Tags.hpp"
-#include "DataStructures/DataBox/Tag.hpp"
 #include "Domain/Creators/RegisterDerivedWithCharm.hpp"
 #include "Domain/Creators/TimeDependence/RegisterDerivedWithCharm.hpp"
 #include "Domain/FunctionsOfTime/RegisterDerivedWithCharm.hpp"
-#include "Domain/Tags.hpp"
-#include "Domain/TagsCharacteresticSpeeds.hpp"
-#include "Evolution/Actions/AddMeshVelocityNonconservative.hpp"
-#include "Evolution/ComputeTags.hpp"
-#include "Evolution/DiscontinuousGalerkin/Actions/ApplyBoundaryCorrections.hpp"
-#include "Evolution/DiscontinuousGalerkin/Actions/ComputeTimeDerivative.hpp"
-#include "Evolution/DiscontinuousGalerkin/DgElementArray.hpp"
-#include "Evolution/DiscontinuousGalerkin/Initialization/Mortars.hpp"
-#include "Evolution/DiscontinuousGalerkin/Initialization/QuadratureTag.hpp"
-#include "Evolution/Initialization/DgDomain.hpp"
-#include "Evolution/Initialization/Evolution.hpp"
-#include "Evolution/Initialization/NonconservativeSystem.hpp"
-#include "Evolution/Initialization/SetVariables.hpp"
-#include "Evolution/NumericInitialData.hpp"
-#include "Evolution/Systems/GeneralizedHarmonic/BoundaryConditions/Factory.hpp"
+#include "Evolution/Executables/GeneralizedHarmonic/GeneralizedHarmonicBase.hpp"
 #include "Evolution/Systems/GeneralizedHarmonic/BoundaryConditions/RegisterDerivedWithCharm.hpp"
-#include "Evolution/Systems/GeneralizedHarmonic/BoundaryCorrections/Factory.hpp"
 #include "Evolution/Systems/GeneralizedHarmonic/BoundaryCorrections/RegisterDerived.hpp"
 #include "Evolution/Systems/GeneralizedHarmonic/ConstraintDamping/RegisterDerivedWithCharm.hpp"
-#include "Evolution/Systems/GeneralizedHarmonic/ConstraintDamping/Tags.hpp"
-#include "Evolution/Systems/GeneralizedHarmonic/Equations.hpp"
-#include "Evolution/Systems/GeneralizedHarmonic/GaugeSourceFunctions/InitializeDampedHarmonic.hpp"
-#include "Evolution/Systems/GeneralizedHarmonic/Initialize.hpp"
-#include "Evolution/Systems/GeneralizedHarmonic/System.hpp"
-#include "Evolution/Systems/GeneralizedHarmonic/Tags.hpp"
-#include "Evolution/Systems/GeneralizedHarmonic/TimeDerivative.hpp"
-#include "Evolution/TypeTraits.hpp"
-#include "IO/Importers/Actions/ReadVolumeData.hpp"
-#include "IO/Importers/Actions/ReceiveVolumeData.hpp"
-#include "IO/Importers/Actions/RegisterWithElementDataReader.hpp"
-#include "IO/Importers/ElementDataReader.hpp"
-#include "IO/Observer/Actions/RegisterEvents.hpp"
-#include "IO/Observer/Helpers.hpp"
-#include "IO/Observer/ObserverComponent.hpp"
-#include "IO/Observer/Tags.hpp"
-#include "NumericalAlgorithms/DiscontinuousGalerkin/Formulation.hpp"
-#include "NumericalAlgorithms/DiscontinuousGalerkin/Tags.hpp"
-#include "NumericalAlgorithms/Interpolation/AddTemporalIdsToInterpolationTarget.hpp"
-#include "NumericalAlgorithms/Interpolation/Callbacks/FindApparentHorizon.hpp"
-#include "NumericalAlgorithms/Interpolation/Callbacks/ObserveTimeSeriesOnSurface.hpp"
-#include "NumericalAlgorithms/Interpolation/CleanUpInterpolator.hpp"
-#include "NumericalAlgorithms/Interpolation/InitializeInterpolationTarget.hpp"
-#include "NumericalAlgorithms/Interpolation/Interpolate.hpp"
-#include "NumericalAlgorithms/Interpolation/InterpolationTarget.hpp"
-#include "NumericalAlgorithms/Interpolation/InterpolationTargetApparentHorizon.hpp"
-#include "NumericalAlgorithms/Interpolation/InterpolationTargetReceiveVars.hpp"
-#include "NumericalAlgorithms/Interpolation/Interpolator.hpp"
-#include "NumericalAlgorithms/Interpolation/InterpolatorReceivePoints.hpp"
-#include "NumericalAlgorithms/Interpolation/InterpolatorReceiveVolumeData.hpp"
-#include "NumericalAlgorithms/Interpolation/InterpolatorRegisterElement.hpp"
-#include "NumericalAlgorithms/Interpolation/Tags.hpp"
-#include "NumericalAlgorithms/Interpolation/TryToInterpolate.hpp"
 #include "Options/Options.hpp"
 #include "Options/Protocols/FactoryCreation.hpp"
-#include "Parallel/Actions/SetupDataBox.hpp"
-#include "Parallel/Actions/TerminatePhase.hpp"
-#include "Parallel/Algorithms/AlgorithmSingleton.hpp"
-#include "Parallel/InitializationFunctions.hpp"
-#include "Parallel/PhaseControl/ExecutePhaseChange.hpp"
 #include "Parallel/PhaseControl/PhaseControlTags.hpp"
-#include "Parallel/PhaseControl/VisitAndReturn.hpp"
-#include "Parallel/PhaseDependentActionList.hpp"
-#include "Parallel/Reduction.hpp"
 #include "Parallel/RegisterDerivedClassesWithCharm.hpp"
-#include "ParallelAlgorithms/Actions/MutateApply.hpp"
-#include "ParallelAlgorithms/Events/Factory.hpp"
-#include "ParallelAlgorithms/EventsAndTriggers/Actions/RunEventsAndTriggers.hpp"
-#include "ParallelAlgorithms/EventsAndTriggers/Completion.hpp"
-#include "ParallelAlgorithms/EventsAndTriggers/Event.hpp"
-#include "ParallelAlgorithms/EventsAndTriggers/EventsAndTriggers.hpp"
-#include "ParallelAlgorithms/EventsAndTriggers/LogicalTriggers.hpp"
-#include "ParallelAlgorithms/EventsAndTriggers/Tags.hpp"
-#include "ParallelAlgorithms/EventsAndTriggers/Trigger.hpp"
-#include "ParallelAlgorithms/Initialization/Actions/AddComputeTags.hpp"
-#include "ParallelAlgorithms/Initialization/Actions/RemoveOptionsAndTerminatePhase.hpp"
-#include "PointwiseFunctions/AnalyticSolutions/GeneralRelativity/KerrSchild.hpp"
-#include "PointwiseFunctions/AnalyticSolutions/GeneralRelativity/WrappedGr.hpp"
-#include "PointwiseFunctions/AnalyticSolutions/Tags.hpp"
-#include "PointwiseFunctions/GeneralRelativity/Christoffel.hpp"
-#include "PointwiseFunctions/GeneralRelativity/DetAndInverseSpatialMetric.hpp"
-#include "PointwiseFunctions/GeneralRelativity/GeneralizedHarmonic/ConstraintGammas.hpp"
-#include "PointwiseFunctions/GeneralRelativity/IndexManipulation.hpp"
-#include "PointwiseFunctions/GeneralRelativity/Ricci.hpp"
-#include "PointwiseFunctions/GeneralRelativity/SpatialMetric.hpp"
-#include "PointwiseFunctions/GeneralRelativity/Tags.hpp"
-#include "Time/Actions/AdvanceTime.hpp"
-#include "Time/Actions/ChangeSlabSize.hpp"
-#include "Time/Actions/ChangeStepSize.hpp"
-#include "Time/Actions/RecordTimeStepperData.hpp"
-#include "Time/Actions/SelfStartActions.hpp"
-#include "Time/Actions/UpdateU.hpp"
-#include "Time/StepChoosers/Factory.hpp"
-#include "Time/StepChoosers/StepChooser.hpp"
+#include "PointwiseFunctions/AnalyticSolutions/GeneralRelativity/GaugeWave.hpp"
 #include "Time/StepControllers/Factory.hpp"
-#include "Time/StepControllers/StepController.hpp"
-#include "Time/Tags.hpp"
-#include "Time/TimeSequence.hpp"
-#include "Time/TimeSteppers/TimeStepper.hpp"
-#include "Time/Triggers/TimeTriggers.hpp"
 #include "Utilities/Blas.hpp"
 #include "Utilities/ErrorHandling/Error.hpp"
 #include "Utilities/ErrorHandling/FloatingPointExceptions.hpp"
-#include "Utilities/Functional.hpp"
-#include "Utilities/MemoryHelpers.hpp"
-#include "Utilities/ProtocolHelpers.hpp"
-#include "Utilities/TMPL.hpp"
-
-/// \cond
-namespace Frame {
-// IWYU pragma: no_forward_declare MathFunction
-struct Inertial;
-}  // namespace Frame
-namespace PUP {
-class er;
-}  // namespace PUP
-namespace Parallel {
-template <typename Metavariables>
-class CProxy_GlobalCache;
-}  // namespace Parallel
-/// \endcond
 
 // First template parameter specifies the source of the initial data, which
 // could be an analytic solution, analytic data, or imported numerical data.
 // Second template parameter specifies the analytic solution used when imposing
 // dirichlet boundary conditions or against which to compute error norms.
 template <typename InitialData, typename BoundaryConditions>
-struct EvolutionMetavars {
-  static constexpr size_t volume_dim = 3;
-  using frame = Frame::Inertial;
-  using system = GeneralizedHarmonic::System<volume_dim>;
-  static constexpr dg::Formulation dg_formulation =
-      dg::Formulation::StrongInertial;
-  static constexpr bool use_damped_harmonic_rollon = true;
-  // Set override_cubic_functions_of_time to true to override the cubic
-  // piecewise polynomial functions of time using
-  // `read_spec_third_order_piecewise_polynomial()`
-  static constexpr bool override_cubic_functions_of_time = false;
-  using temporal_id = Tags::TimeStepId;
-  static constexpr bool local_time_stepping = false;
-  using initial_data = InitialData;
-  using analytic_solution_tag = Tags::AnalyticSolution<BoundaryConditions>;
-
-  using time_stepper_tag = Tags::TimeStepper<
-      tmpl::conditional_t<local_time_stepping, LtsTimeStepper, TimeStepper>>;
-
-  using analytic_solution_fields = typename system::variables_tag::tags_list;
-  using observe_fields = tmpl::append<
-      tmpl::push_back<
-          analytic_solution_fields,
-          ::Tags::PointwiseL2Norm<
-              GeneralizedHarmonic::Tags::GaugeConstraint<volume_dim, frame>>,
-          ::Tags::PointwiseL2Norm<GeneralizedHarmonic::Tags::
-                                      ThreeIndexConstraint<volume_dim, frame>>>,
-      tmpl::conditional_t<volume_dim == 3,
-                          tmpl::list<::Tags::PointwiseL2Norm<
-                              GeneralizedHarmonic::Tags::FourIndexConstraint<
-                                  volume_dim, frame>>>,
-                          tmpl::list<>>>;
-
-  struct AhA {
-    using tags_to_observe =
-        tmpl::list<StrahlkorperGr::Tags::AreaCompute<frame>>;
-    using compute_items_on_source = tmpl::list<
-        gr::Tags::SpatialMetricCompute<volume_dim, frame, DataVector>,
-        ah::Tags::InverseSpatialMetricCompute<volume_dim, frame>,
-        ah::Tags::ExtrinsicCurvatureCompute<volume_dim, frame>,
-        ah::Tags::SpatialChristoffelSecondKindCompute<volume_dim, frame>>;
-    using vars_to_interpolate_to_target =
-        tmpl::list<gr::Tags::SpatialMetric<volume_dim, frame, DataVector>,
-                   gr::Tags::InverseSpatialMetric<volume_dim, frame>,
-                   gr::Tags::ExtrinsicCurvature<volume_dim, frame>,
-                   gr::Tags::SpatialChristoffelSecondKind<volume_dim, frame>>;
-    using compute_items_on_target = tmpl::append<
-        tmpl::list<StrahlkorperGr::Tags::AreaElementCompute<frame>>,
-        tags_to_observe>;
-    using compute_target_points =
-        intrp::TargetPoints::ApparentHorizon<AhA, ::Frame::Inertial>;
-    using post_interpolation_callback =
-        intrp::callbacks::FindApparentHorizon<AhA, ::Frame::Inertial>;
-    using post_horizon_find_callback =
-        intrp::callbacks::ObserveTimeSeriesOnSurface<tags_to_observe, AhA, AhA>;
-  };
-  using interpolation_target_tags = tmpl::list<AhA>;
-  using interpolator_source_vars =
-      tmpl::list<gr::Tags::SpacetimeMetric<volume_dim, frame>,
-                 GeneralizedHarmonic::Tags::Pi<volume_dim, frame>,
-                 GeneralizedHarmonic::Tags::Phi<volume_dim, frame>>;
-
-  struct factory_creation
-      : tt::ConformsTo<Options::protocols::FactoryCreation> {
-    using factory_classes = tmpl::map<
-        tmpl::pair<
-            Event,
-            tmpl::flatten<tmpl::list<
-                Events::Completion,
-                dg::Events::field_observations<volume_dim, Tags::Time,
-                                               observe_fields,
-                                               analytic_solution_fields>,
-                Events::time_events<system>,
-                intrp::Events::Interpolate<3, AhA, interpolator_source_vars>>>>,
-        tmpl::pair<StepChooser<StepChooserUse::LtsStep>,
-                   StepChoosers::standard_step_choosers<system>>,
-        tmpl::pair<StepChooser<StepChooserUse::Slab>,
-                   StepChoosers::standard_slab_choosers<system,
-                                                        local_time_stepping>>,
-        tmpl::pair<StepController, StepControllers::standard_step_controllers>,
-        tmpl::pair<Trigger, tmpl::append<Triggers::logical_triggers,
-                                         Triggers::time_triggers>>>;
-  };
-
+struct EvolutionMetavars
+    : public virtual GeneralizedHarmonicDefaults,
+      public GeneralizedHarmonicTemplateBase<
+          EvolutionMetavars<InitialData, BoundaryConditions>> {
+  using const_global_cache_tags =
+      typename GeneralizedHarmonicTemplateBase<EvolutionMetavars<
+          InitialData, BoundaryConditions>>::const_global_cache_tags;
   using observed_reduction_data_tags =
-      observers::collect_reduction_data_tags<tmpl::push_back<
-          tmpl::at<typename factory_creation::factory_classes, Event>,
-          typename AhA::post_horizon_find_callback>>;
-
-  using step_actions = tmpl::list<
-      evolution::dg::Actions::ComputeTimeDerivative<EvolutionMetavars>,
-      evolution::dg::Actions::ApplyBoundaryCorrections<EvolutionMetavars>,
-      tmpl::conditional_t<
-          local_time_stepping, tmpl::list<>,
-          tmpl::list<Actions::RecordTimeStepperData<>, Actions::UpdateU<>>>>;
-
-  enum class Phase {
-    Initialization,
-    RegisterWithElementDataReader,
-    ImportInitialData,
-    InitializeInitialDataDependentQuantities,
-    InitializeTimeStepperHistory,
-    Register,
-    LoadBalancing,
-    Evolve,
-    Exit
-  };
-
-  static std::string phase_name(Phase phase) noexcept {
-    if (phase == Phase::LoadBalancing) {
-      return "LoadBalancing";
-    }
-    ERROR(
-        "Passed phase that should not be used in input file. Integer "
-        "corresponding to phase is: "
-        << static_cast<int>(phase));
-  }
-
-  using phase_changes = tmpl::list<PhaseControl::Registrars::VisitAndReturn<
-      EvolutionMetavars, Phase::LoadBalancing>>;
-
-  using initialize_phase_change_decision_data =
-      PhaseControl::InitializePhaseChangeDecisionData<phase_changes>;
-
-  using phase_change_tags_and_combines_list =
-      PhaseControl::get_phase_change_tags<phase_changes>;
-
-  using const_global_cache_tags = tmpl::list<
-      analytic_solution_tag, time_stepper_tag, Tags::EventsAndTriggers,
-      GeneralizedHarmonic::ConstraintDamping::Tags::DampingFunctionGamma0<
-          volume_dim, frame>,
-      GeneralizedHarmonic::ConstraintDamping::Tags::DampingFunctionGamma1<
-          volume_dim, frame>,
-      GeneralizedHarmonic::ConstraintDamping::Tags::DampingFunctionGamma2<
-          volume_dim, frame>,
-      PhaseControl::Tags::PhaseChangeAndTriggers<phase_changes>>;
-
-  using dg_registration_list =
-      tmpl::list<intrp::Actions::RegisterElementWithInterpolator,
-                 observers::Actions::RegisterEventsWithObservers>;
-
-  using initialization_actions = tmpl::list<
-      Actions::SetupDataBox,
-      Initialization::Actions::TimeAndTimeStep<EvolutionMetavars>,
-      evolution::dg::Initialization::Domain<volume_dim,
-                                            override_cubic_functions_of_time>,
-      Initialization::Actions::NonconservativeSystem<system>,
-      tmpl::conditional_t<
-          evolution::is_numeric_initial_data_v<initial_data>, tmpl::list<>,
-          evolution::Initialization::Actions::SetVariables<
-              domain::Tags::Coordinates<volume_dim, Frame::Logical>>>,
-      Initialization::Actions::TimeStepperHistory<EvolutionMetavars>,
-      GeneralizedHarmonic::Actions::InitializeGhAnd3Plus1Variables<volume_dim>,
-      Initialization::Actions::AddComputeTags<tmpl::push_back<
-          StepChoosers::step_chooser_compute_tags<EvolutionMetavars>,
-          evolution::Tags::AnalyticCompute<volume_dim, analytic_solution_tag,
-                                           analytic_solution_fields>>>,
-      ::evolution::dg::Initialization::Mortars<volume_dim, system>,
-      Initialization::Actions::RemoveOptionsAndTerminatePhase>;
-
-  using initialize_initial_data_dependent_quantities_actions = tmpl::list<
-      GeneralizedHarmonic::gauges::Actions::InitializeDampedHarmonic<
-          volume_dim, use_damped_harmonic_rollon>,
-      GeneralizedHarmonic::Actions::InitializeConstraints<volume_dim>,
-      Parallel::Actions::TerminatePhase>;
-
-  using dg_element_array = DgElementArray<
-      EvolutionMetavars,
-      tmpl::flatten<tmpl::list<
-          Parallel::PhaseActions<Phase, Phase::Initialization,
-                                 initialization_actions>,
-          tmpl::conditional_t<
-              evolution::is_numeric_initial_data_v<initial_data>,
-              tmpl::list<
-                  Parallel::PhaseActions<
-                      Phase, Phase::RegisterWithElementDataReader,
-                      tmpl::list<
-                          importers::Actions::RegisterWithElementDataReader,
-                          Parallel::Actions::TerminatePhase>>,
-                  Parallel::PhaseActions<
-                      Phase, Phase::ImportInitialData,
-                      tmpl::list<importers::Actions::ReadVolumeData<
-                                     evolution::OptionTags::NumericInitialData,
-                                     typename system::variables_tag::tags_list>,
-                                 importers::Actions::ReceiveVolumeData<
-                                     evolution::OptionTags::NumericInitialData,
-                                     typename system::variables_tag::tags_list>,
-                                 Parallel::Actions::TerminatePhase>>>,
-              tmpl::list<>>,
-          Parallel::PhaseActions<
-              Phase, Phase::InitializeInitialDataDependentQuantities,
-              initialize_initial_data_dependent_quantities_actions>,
-          Parallel::PhaseActions<
-              Phase, Phase::InitializeTimeStepperHistory,
-              SelfStart::self_start_procedure<step_actions, system>>,
-          Parallel::PhaseActions<Phase, Phase::Register,
-                                 tmpl::list<dg_registration_list,
-                                            Parallel::Actions::TerminatePhase>>,
-          Parallel::PhaseActions<
-              Phase, Phase::Evolve,
-              tmpl::list<Actions::RunEventsAndTriggers, Actions::ChangeSlabSize,
-                         step_actions, Actions::AdvanceTime,
-                         PhaseControl::Actions::ExecutePhaseChange<
-                             phase_changes>>>>>>;
-
+      typename GeneralizedHarmonicTemplateBase<EvolutionMetavars<
+          InitialData, BoundaryConditions>>::observed_reduction_data_tags;
+  using component_list = typename GeneralizedHarmonicTemplateBase<
+      EvolutionMetavars<InitialData, BoundaryConditions>>::component_list;
   template <typename ParallelComponent>
-  struct registration_list {
-    using type = std::conditional_t<
-        std::is_same_v<ParallelComponent, dg_element_array>,
-      dg_registration_list, tmpl::list<>>;
-  };
-
-  using component_list = tmpl::flatten<tmpl::list<
-      observers::Observer<EvolutionMetavars>,
-      observers::ObserverWriter<EvolutionMetavars>,
-      intrp::Interpolator<EvolutionMetavars>,
-      intrp::InterpolationTarget<EvolutionMetavars, AhA>,
-      tmpl::conditional_t<evolution::is_numeric_initial_data_v<initial_data>,
-                          importers::ElementDataReader<EvolutionMetavars>,
-                          tmpl::list<>>,
-      dg_element_array>>;
+  using registration_list = typename GeneralizedHarmonicTemplateBase<
+      EvolutionMetavars<InitialData, BoundaryConditions>>::
+      template registration_list<ParallelComponent>;
 
   static constexpr Options::String help{
-      "Evolve a generalized harmonic analytic solution.\n\n"
-      "The analytic solution is: KerrSchild\n"
-      "The numerical flux is:    UpwindFlux\n"};
-
-  template <typename... Tags>
-  static Phase determine_next_phase(
-      const gsl::not_null<tuples::TaggedTuple<Tags...>*>
-          phase_change_decision_data,
-      const Phase& current_phase,
-      const Parallel::CProxy_GlobalCache<EvolutionMetavars>&
-          cache_proxy) noexcept {
-    const auto next_phase =
-        PhaseControl::arbitrate_phase_change<phase_changes>(
-            phase_change_decision_data, current_phase,
-            *(cache_proxy.ckLocalBranch()));
-    if (next_phase.has_value()) {
-      return next_phase.value();
-    }
-    switch (current_phase) {
-      case Phase::Initialization:
-        return evolution::is_numeric_initial_data_v<initial_data>
-                   ? Phase::RegisterWithElementDataReader
-                   : Phase::InitializeInitialDataDependentQuantities;
-      case Phase::RegisterWithElementDataReader:
-        return Phase::ImportInitialData;
-      case Phase::ImportInitialData:
-        return Phase::InitializeInitialDataDependentQuantities;
-      case Phase::InitializeInitialDataDependentQuantities:
-        return Phase::InitializeTimeStepperHistory;
-      case Phase::InitializeTimeStepperHistory:
-        return Phase::Register;
-      case Phase::Register:
-        return Phase::Evolve;
-      case Phase::Evolve:
-        return Phase::Exit;
-      case Phase::Exit:
-        ERROR(
-            "Should never call determine_next_phase with the current phase "
-            "being 'Exit'");
-      default:
-        ERROR(
-            "Unknown type of phase. Did you static_cast<Phase> an integral "
-            "value?");
-    }
-  }
-
-  // NOLINTNEXTLINE(google-runtime-references)
-  void pup(PUP::er& /*p*/) noexcept {}
+      "Evolve the Einstein field equations using the Generalized Harmonic "
+      "formulation\n"};
 };
 
 static const std::vector<void (*)()> charm_init_node_funcs{
